improved test for reading cmlh5 to xarray
